refactor(clients): 🔨 extract retrieval, renewal, import into partial files
